% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixar_films}
\alias{pixar_films}
\title{Pixar films}
\format{
A data frame with 27 rows and 5 variables:
\tabular{ll}{
\code{number} \tab Order of release. \cr
\code{film} \tab Film title. \cr
\code{release_date} \tab Date the film premiered. \cr
\code{run_time} \tab Film length in minutes. \cr
\code{film_rating} \tab MPA film rating.
}
}
\source{
TidyTuesday, 2025-03-11.
}
\usage{
pixar_films
}
\description{
Basic information about Pixar feature films, including release order,
release date, runtime, and film rating.
}
\keyword{datasets}
