% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_response}
\alias{public_response}
\title{Pixar public response}
\format{
A data frame with 24 rows and 5 variables:
\tabular{ll}{
\code{film} \tab Film title. \cr
\code{rotten_tomatoes} \tab Rotten Tomatoes score out of 100. \cr
\code{metacritic} \tab Metacritic score out of 100. \cr
\code{cinema_score} \tab CinemaScore letter grade. \cr
\code{critics_choice} \tab Critics Choice score out of 100.
}
}
\source{
TidyTuesday, 2025-03-11.
}
\usage{
public_response
}
\description{
Public and critical response measures for Pixar films.
}
\keyword{datasets}
